audio/serializers.py
from rest_framework import serializers
from .models import Group, Audio


class GroupSerializer(serializers.ModelSerializer):
    class Meta:
        model = Group
        fields = ('id', 'name', 'is_available')


class AudioSerializer(serializers.ModelSerializer):
    class Meta:
        model = Audio
        fields = ('id', 'category', 'description', 'source', 'is_available')