requirements.txt
appdirs==1.4.0
cffi==1.9.1
cryptography==1.7.2
defusedxml==0.5.0rc1
Django==1.10.5
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-cors-headers==2.0.2
django-jet==1.0.4
django-user-accounts==2.0.0
djangorestframework==3.5.3
enum34==1.1.6
gunicorn==19.6.0
idna==2.2
ipaddress==1.0.18
jsonfield==1.0.3
oauthlib==2.0.1
packaging==16.8
paypalrestsdk==1.12.0
pinax-eventlog==1.1.2
pinax-theme-bootstrap==7.11.0
pinax-webanalytics==2.0.4
pkg-resources==0.0.0
psycopg2==2.6.2
pyasn1==0.2.2
pycparser==2.17
PyJWT==1.4.2
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-openid==2.2.5
python-social-auth==0.2.13
python3-openid==3.0.10
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.7.0
six==1.10.0
social-auth-app-django==1.0.1
social-auth-core==1.1.0
whitenoise==3.3.0
